+++ esp-hal/src/lib.rs
@@ -255,14 +255,15 @@ mod macros;
 
 #[cfg(feature = "rt")]
 pub use procmacros::blocking_main as main;
-#[cfg(any(lp_core, ulp_riscv_core))]
-#[cfg(feature = "unstable")]
-#[cfg_attr(docsrs, doc(cfg(feature = "unstable")))]
-pub use procmacros::load_lp_code;
 #[cfg_attr(docsrs, doc(cfg(feature = "unstable")))]
 #[instability::unstable]
 #[cfg_attr(not(feature = "unstable"), allow(unused))]
-pub use procmacros::{handler, ram};
+pub use procmacros::handler;
+#[cfg(any(lp_core, ulp_riscv_core))]
+#[instability::unstable]
+#[cfg_attr(docsrs, doc(cfg(feature = "unstable")))]
+pub use procmacros::load_lp_code;
+pub use procmacros::ram;
 
 #[cfg(all(feature = "rt", feature = "exception-handler"))]
 mod exception_handler;